add padding mask
